Tyan S8226: Fix incompatible pointer warning
Fix warning:
  mptable.c:52, GNU Compiler 4 (gcc), Priority: Normal
  passing argument 3 of 'mptable_write_buses' from incompatible pointer type [enabled by default]

mptable_write_buses is expecting a pointer to an int, so I changed the
U8 isa_bus to an int to match.  A U8 doesn't make sense if the value could
be greater than 255 - certainly unlikely, but possible since the value
of isa_bus gets set to the maximum PCI bus number + 1.

Reviewed-on: http://review.coreboot.org/2450
Tested-by: build bot (Jenkins)

+++ src/mainboard/tyan/s8226/get_bus_conf.c
@@ -32,7 +32,7 @@
 /* Global variables for MB layouts and these will be shared by irqtable mptable
  * and acpi_tables busnum is default.
  */
-u8 bus_isa;
+int bus_isa;
 u8 bus_sp5100[2];
 u8 bus_sr5650[14];
 
@@ -101,7 +101,7 @@ void get_bus_conf(void)
 	if (dev) {
 		bus_sp5100[1] = pci_read_config8(dev, PCI_SECONDARY_BUS);
 
-		bus_isa = pci_read_config8(dev, PCI_SUBORDINATE_BUS);
+		bus_isa = (int)pci_read_config8(dev, PCI_SUBORDINATE_BUS);
 		bus_isa++;
 		for (j = bus_sp5100[1]; j < bus_isa; j++)
 			bus_type[j] = 1;
@@ -113,7 +113,7 @@ void get_bus_conf(void)
 		if (dev) {
 			bus_sr5650[i] = pci_read_config8(dev, PCI_SECONDARY_BUS);
 			if(255 != bus_sr5650[i]) {
-				bus_isa = pci_read_config8(dev, PCI_SUBORDINATE_BUS);
+				bus_isa = (int)pci_read_config8(dev, PCI_SUBORDINATE_BUS);
 				bus_isa++;
 				bus_type[bus_sr5650[i]] = 1; /* PCI bus. */
 			}
